Add files via upload
implementing final logic about buying Pocket Tokens, implementing buying cosmetics, and refunding cosmetics using Pocket Tokens.

+++ Pocket Entertainment/PocketEntertainmentApp/src/main/java/com/pockEtentertainmentApp/game/model/Game.java
@@ -40,6 +40,6 @@ public class Game {
     private User creator;
 
     @Column(nullable = false)
-    private int downloads;
+    private long downloads;
 
 }
